srcs/backend/requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
djangorestframework-simplejwt
django-cors-headers